Create a build phase
