+++ Generator/Template.swift
@@ -1,0 +1,7 @@
+//
+//  {filename}
+//  Decodable
+//
+//  Generated automatically by {by} as a build phase.
+//  Copyright © 2016 anviking. All rights reserved.
+//
